LayoutTests/imported/w3c/web-platform-tests/intersection-observer/shadow-content-expected.txt


PASS Observing a target inside shadow DOM. 
FAIL First rAF after creating shadow DOM. assert_equals: entries[0].boundingClientRect.left expected 8 but got 0
